requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
pydantic-settings==2.2.1
SQLAlchemy==2.0.30
asyncpg==0.29.0
python-multipart==0.0.20
python-decouple==3.8
psycopg2-binary==2.9.10
pandas==2.3.1
openpyxl==3.1.5

# === Embeddings / IA ===
# openai==1.30.1
langchain==0.2.7
langchain_openai==0.1.16
pinecone-client==4.1.2
langchain_openai==0.1.16
alembic==1.16.4


# === Terminal / CLI (agente) ===
# rich==13.7.1
